src/requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
openpyxl>=3.1.0

# Backend API dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6